remove compile warnings

+++ source/DataElement/AMDSDwellStatusData.cpp
@@ -1,7 +1,7 @@
 #include "AMDSDwellStatusData.h"
 
 /// =============== implementation of AMDSStatusData ================
-AMDSDwellStatusData::AMDSDwellStatusData(quint32 fastCounts, quint32 slowCounts, quint64 detectorTemperature, quint64 accumulationTime, quint64 liveTime, quint64 realTime, quint32 generalPurposeCounter, const QTime &dwellStartTime, const QTime &dwellEndTime, const QTime &dwellReplyTime)
+AMDSDwellStatusData::AMDSDwellStatusData(int fastCounts, int slowCounts, double detectorTemperature, double accumulationTime, double liveTime, double realTime, int generalPurposeCounter, const QTime &dwellStartTime, const QTime &dwellEndTime, const QTime &dwellReplyTime)
 {
 	setStatusData(fastCounts, slowCounts, detectorTemperature, accumulationTime, liveTime, realTime, generalPurposeCounter, dwellStartTime, dwellEndTime, dwellReplyTime);
 }
@@ -18,13 +18,13 @@ AMDSDwellStatusData::~AMDSDwellStatusData()
 
 AMDSDwellStatusData& AMDSDwellStatusData::operator +(const AMDSDwellStatusData &other)
 {
-	quint32 fastCountsPlus = fastCounts() + other.fastCounts();
-	quint32 slowCountsPlus = slowCounts() + other.slowCounts();
-	quint64 detectorTemperaturePlus = detectorTemperature() + other.detectorTemperature();
-	quint64 accumulationTimePlus = accumulationTime() + other.accumulationTime();
-	quint64 liveTimePlus = liveTime() + other.liveTime();
-	quint64 realTimePlus = realTime() + other.realTime();
-	quint32 generalPurposeCounterPlus = generalPurposeCounter() + other.generalPurposeCounter();
+	int fastCountsPlus = fastCounts() + other.fastCounts();
+	int slowCountsPlus = slowCounts() + other.slowCounts();
+	double detectorTemperaturePlus = detectorTemperature() + other.detectorTemperature();
+	double accumulationTimePlus = accumulationTime() + other.accumulationTime();
+	double liveTimePlus = liveTime() + other.liveTime();
+	double realTimePlus = realTime() + other.realTime();
+	int generalPurposeCounterPlus = generalPurposeCounter() + other.generalPurposeCounter();
 
 	QTime dwellStartTimePlus =  dwellStartTime();
 	QTime dwellEndTimePlus =  dwellEndTime();
@@ -39,15 +39,15 @@ AMDSDwellStatusData& AMDSDwellStatusData::operator +(const AMDSDwellStatusData &
 	return *this;
 }
 
-AMDSDwellStatusData& AMDSDwellStatusData::operator /(quint32 divisor)
+AMDSDwellStatusData& AMDSDwellStatusData::operator /(int divisor)
 {
-	quint32 fastCountsDiv = fastCounts() / divisor;
-	quint32 slowCountsDiv = slowCounts() / divisor;
-	quint64 detectorTemperatureDiv = detectorTemperature() / divisor;
-	quint64 accumulationTimeDiv = accumulationTime() / divisor;
-	quint64 liveTimeDiv = liveTime() / divisor;
-	quint64 realTimeDiv = realTime() / divisor;
-	quint32 generalPurposeCounterDiv = generalPurposeCounter() / divisor;
+	int fastCountsDiv = fastCounts() / divisor;
+	int slowCountsDiv = slowCounts() / divisor;
+	double detectorTemperatureDiv = detectorTemperature() / divisor;
+	double accumulationTimeDiv = accumulationTime() / divisor;
+	double liveTimeDiv = liveTime() / divisor;
+	double realTimeDiv = realTime() / divisor;
+	int generalPurposeCounterDiv = generalPurposeCounter() / divisor;
 
 	setStatusData(fastCountsDiv, slowCountsDiv, detectorTemperatureDiv, accumulationTimeDiv, liveTimeDiv, realTimeDiv, generalPurposeCounterDiv, dwellStartTime(), dwellEndTime(), dwellReplyTime());
 	return *this;
@@ -156,7 +156,7 @@ QString AMDSDwellStatusData::toString() const
 	return QString("Status: %1 %2 %3 %4").arg(dwellStartTime_.toString()).arg(dwellEndTime_.toString()).arg(dwellReplyTime_.toString()).arg(liveTime_);
 }
 
-void AMDSDwellStatusData::setStatusData(quint32 fastCounts, quint32 slowCounts, quint64 detectorTemperature, quint64 accumulationTime, quint64 liveTime, quint64 realTime, quint32 generalPurposeCounter, const QTime &dwellStartTime, const QTime &dwellEndTime, const QTime &dwellReplyTime)
+void AMDSDwellStatusData::setStatusData(int fastCounts, int slowCounts, double detectorTemperature, double accumulationTime, double liveTime, double realTime, int generalPurposeCounter, const QTime &dwellStartTime, const QTime &dwellEndTime, const QTime &dwellReplyTime)
 {
 	fastCounts_ = fastCounts ;
 	slowCounts_ = slowCounts ;
